# api/openapi_librebread.yml
openapi: 3.0.0
info:
  title: LibreBread API
  version: 0.0.2
paths:
  /:
    get:
      responses:
        '302':
          description: Redirect to sms handler
          headers:
            Location:
              schema:
                type: string
  /sms:
    get:
      summary: Returns a list of SMS.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int64
            default: 50
          description: The number of latest sms
        - in: query
          name: json
          schema:
            type: boolean
          description: Should be JSON response
      responses:
        '200':
          description: A array of SMS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SMSList'
            text/html:
              schema:
                type: string
        '500':
          description: Internal error ocured
          content:
            text/plain:
              schema:
                type: string
  /helpdesk:
    get:
      operationId: GetHelpdeskEddyTicket
      summary: List tickets
      description: List tickets
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            format: int64
            default: 50
          description: The number of latest sms
        - in: query
          name: json
          schema:
            type: boolean
          description: Should be JSON response
      responses:
        '200':
          description: A array of tickets (default last 50)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelpdeskEddyTicketList'
            text/html:
              schema:
                type: string
        '500':
          description: Internal error ocured
          content:
            text/plain:
              schema:
                type: string
  /libre/send:
    post:
      summary: Send the fake SMS
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - from
                - to
                - text
              properties:
                from:
                  type: string
                to:
                  type: string
                text:
                  type: string
      responses:
        '200':
          description: List of message ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LibreBreadSMSIds'
        '400':
          description: Some filed is missed
        '500':
          description: Internal error ocured
          content:
            text/plain:
              schema:
                type: string
  /libre/check:
    post:
      description: Not implemented
      responses:
        '501':
          description: Not implemented
  /api/v2/tickets:    
    post:
      operationId: PostHelpdeskEddyTicket
      summary: Create HelpdeskEddy ticket
      description: Create HelpdeskEddy ticket
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - title
                - description
              properties:
                type_id:
                  type: integer
                priority_id:
                  type: integer
                department_id:
                  type: integer
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Ticket succesfuly created
        '400':
          description: One of required fileds is missed
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal error ocured
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    SMS:
      type: object
      required:
        - ID
        - Time
        - From
        - To
        - Text
        - Provider
      properties:
        ID:
          type: string
        Time:
          type: string
          format: date-time
        From:
          type: string
        To:
          type: string
        Text:
          type: string
        Provider:
          type: string
    SMSList:
      type: array
      items:
        $ref: '#/components/schemas/SMS'
    LibreBreadSMSIds:
      type: array
      items:
        type: string
    HelpdeskEddyTicket:
      type: object
      required:
        - ID
        - Time
        - TypeId
        - PriorityId
        - DepartmentId
        - Title
        - Description
      properties:
        ID:
          type: string
        Time:
          type: string
          format: date-time
        TypeId:
          type: integer
          minimum : 0
        PriorityId:
          type: integer
          minimum : 0
        DepartmentId:
          type: integer
          minimum : 0
        Title:
          type: string
        Description:
          type: string
    HelpdeskEddyTicketList:
      type: array
      items:
        $ref: '#/components/schemas/HelpdeskEddyTicket'